# docs/swagger.yaml
basePath: /api
definitions:
  db.Day:
    properties:
      day:
        type: integer
      id:
        type: integer
      month:
        type: integer
      saldo:
        description: Saldo is DailyBudget - Cost of all Spends. It can be negative
        type: number
      spends:
        items:
          $ref: '#/definitions/db.Spend'
        type: array
      year:
        type: integer
    type: object
...
  db.Spend:
    properties:
      cost:
        type: number
      day:
        type: integer
      id:
        type: integer
      month:
        type: integer
      notes:
        type: string
      title:
        type: string
      type:
        $ref: '#/definitions/db.SpendType'
      year:
        type: integer
    type: object
  db.SpendType:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.GetDayResp:
    properties:
      day:
        $ref: '#/definitions/db.Day'
      error:
        description: Error is specified only when success if false
        type: string
      request_id:
        type: string
      success:
        type: boolean
    type: object
...
  models.Response:
    properties:
      error:
        description: Error is specified only when success if false
        type: string
      request_id:
        type: string
      success:
        type: boolean
    type: object
  models.SearchSpendsResp:
    properties:
      error:
        description: Error is specified only when success if false
        type: string
      request_id:
        type: string
      spends:
        items:
          $ref: '#/definitions/db.Spend'
        type: array
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: Easy-to-use, lightweight and self-hosted solution to track your finances - [GitHub](https://github.com/ShoshinNikita/budget-manager)
  license:
    name: MIT
    url: https://github.com/ShoshinNikita/budget-manager/blob/master/LICENSE
  title: Budget Manager API
  version: v0.2
paths:
  /api/days/date:
    get:
      parameters:
      - in: query
        name: day
        required: true
        type: integer
      - in: query
        name: month
        required: true
        type: integer
      - in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetDayResp'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Day doesn't exist
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get Day by date
      tags:
      - Days
  /api/days/id:
    get:
      parameters:
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetDayResp'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Day doesn't exist
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get Day by id
      tags:
      - Days
...
  /api/search/spends:
    get:
      parameters:
      - description: After must be in the RFC3339 format (https://tools.ietf.org/html/rfc3339#section-5.8)
        format: date
        in: query
        name: after
        type: string
      - description: Before must be in the RFC3339 format (https://tools.ietf.org/html/rfc3339#section-5.8)
        format: date
        in: query
        name: before
        type: string
      - in: query
        name: max_cost
        type: number
      - in: query
        name: min_cost
        type: number
      - description: Notes can be in any case. Search will be performed by lowercased value
        in: query
        name: notes
        type: string
      - description: NotesExactly defines should we search exactly for the given notes
        in: query
        name: notes_exactly
        type: boolean
      - description: 'Order specify sort order. Available options: asc (default), desc'
        in: query
        name: order
        type: string
      - description: 'Sort specify field to sort by. Available options: title, cost, date (default)'
        in: query
        name: sort
        type: string
      - description: Title can be in any case. Search will be performed by lowercased value
        in: query
        name: title
        type: string
      - description: TitleExactly defines should we search exactly for the given title
        in: query
        name: title_exactly
        type: boolean
      - description: TypeIDs is a list of Spend Type ids to search for
        in: query
        items:
          type: integer
        name: type_ids
        type: array
      - description: WithoutType is used to search for Spends without Spend Type. TypeIDs are ignored when it is true
        in: query
        name: without_type
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SearchSpendsResp'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Search Spends
      tags:
      - Search
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"